moved TestCurrencyProvider into proper location
